+++ build.zig
@@ -662,7 +662,7 @@ pub fn build(b: *std.Build) void {
     cli_version.addArg("version");
     cli_version.expectStdOutEqual(
         "caudex 0.1.0\nengine: 0.1.0 (schema 1)\npersistence contract: 2\n" ++
-            "tracking contract: 6\nsqlite adapter: 0.1.0 (schema 1-7)\n",
+            "tracking contract: 6\nsqlite adapter: 0.1.0 (schema 1-8)\n",
     );
 
     const cli_database_human = b.addRunArtifact(cli);
@@ -671,8 +671,8 @@ pub fn build(b: *std.Build) void {
         \\Database: :memory:
         \\Kind: memory
         \\Adapter version: 0.1.0
-        \\Schema version: 7
-        \\Supported schema: 1-7
+        \\Schema version: 8
+        \\Supported schema: 1-8
         \\Compatibility: current
         \\
     );
@@ -689,8 +689,8 @@ pub fn build(b: *std.Build) void {
     cli_database_json.expectStdOutEqual(
         "{\"schemaVersion\":1,\"kind\":\"caudex.database.info\",\"data\":{" ++
             "\"databasePath\":\":memory:\",\"databaseKind\":\"memory\"," ++
-            "\"adapterVersion\":\"0.1.0\",\"databaseSchemaVersion\":7," ++
-            "\"minimumSchemaVersion\":1,\"latestSchemaVersion\":7," ++
+            "\"adapterVersion\":\"0.1.0\",\"databaseSchemaVersion\":8," ++
+            "\"minimumSchemaVersion\":1,\"latestSchemaVersion\":8," ++
             "\"compatibility\":\"current\"}}\n",
     );
 
@@ -735,8 +735,8 @@ pub fn build(b: *std.Build) void {
         "{\"schemaVersion\":1,\"kind\":\"caudex.database.info\",\"data\":{" ++
             "\"databasePath\":\".zig-cache/cwe112-broken-pipe.sqlite\"," ++
             "\"databaseKind\":\"file\",\"adapterVersion\":\"0.1.0\"," ++
-            "\"databaseSchemaVersion\":7,\"minimumSchemaVersion\":1," ++
-            "\"latestSchemaVersion\":7,\"compatibility\":\"current\"}}\n",
+            "\"databaseSchemaVersion\":8,\"minimumSchemaVersion\":1," ++
+            "\"latestSchemaVersion\":8,\"compatibility\":\"current\"}}\n",
     );
 
     const cli_workout_start_test = b.addSystemCommand(&.{